+++ applications/qnet-mobile/index.js
@@ -3,6 +3,8 @@
  */
 
 import 'react-native-get-random-values';
+import { Buffer } from 'buffer';
+global.Buffer = Buffer;
 import { AppRegistry } from 'react-native';
 import App from './App';
 import { name as appName } from './app.json';