+++ applications/qnet-mobile/metro.config.js
@@ -8,10 +8,10 @@ const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
  */
 const config = {
   resolver: {
-    alias: {
-      'crypto': 'react-native-crypto-js',
-      'stream': 'stream-browserify',
-      'buffer': '@craftzdog/react-native-buffer',
+    extraNodeModules: {
+      crypto: require.resolve('react-native-crypto'),
+      stream: require.resolve('readable-stream'),
+      events: require.resolve('events'),
     },
   },
 };